backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
gunicorn>=23.0
requests>=2.31
whitenoise>=6.6
